Python Game
This is the First Basic Python Number game that I made while learning the language. So Basically it takes the input from the user as an integer and check it with the  number which is randomly selected by the computer and then compares it and guides the user by giving the output as increase or decrease the value but not by how much and if the user entered value and the computer gets equal then the user will win as he guessed right and also shows the number of trials user took for getting same value!